Relax requirement for URI on Resource Field.
The URI check is just `value =~ /^[a-z]+:/` which doesn't even make sure
the value is a valid ARN just that it starts with a letter and has a
`:`.

We use terraform to manage our StepFunctions. Because we operate in a
few different regions we template out the JSON (still valid json!) an
then use terraform templating to create the template and manage teht
SFN. There is no time after terraform interpolates variables for us to
run a file through the linter (its all in memory).

We are just running into this because instead of interpolating region in
the ARN we are now using a data resource to allow terraform to drop the
whole ARN in.

That is we are going from:
```
"Resource": "arn:aws:lambda:${region}:${account}:function:foobar:${alias}"
```
To:
```
"Resource": "${lambda_arn}"
```

The first passes the very lax regex the second doesn't